+++ src/player.cpp
@@ -210,7 +210,7 @@ Player::~Player()
 	setEditHouse(nullptr);
 }
 
-bool Player::setVocation(uint32_t vocId)
+bool Player::setVocation(uint16_t vocId)
 {
 	Vocation* voc = g_vocations.getVocation(vocId);
 	if (!voc) {
@@ -449,7 +449,7 @@ int32_t Player::getArmor() const
 			armor += inventoryItem->getArmor();
 		}
 	}
-	return int32_t(armor * vocation->armorMultiplier);
+	return static_cast<int32_t>(armor * vocation->armorMultiplier);
 }
 
 void Player::getShieldAndWeapon(const Item*& shield, const Item*& weapon) const
@@ -508,8 +508,8 @@ int32_t Player::getDefense() const
 		return 0;
 	}
 
-	defenseValue = int32_t(defenseValue * vocation->defenseMultiplier);
-	return ((int32_t)std::ceil(((float)(defenseSkill * (defenseValue * 0.015)) + (defenseValue * 0.1)) * defenseFactor));
+	defenseValue = static_cast<int32_t>(defenseValue * vocation->defenseMultiplier);
+	return static_cast<int32_t>(std::ceil(((float)(defenseSkill * (defenseValue * 0.015)) + (defenseValue * 0.1)) * defenseFactor));
 }
 
 float Player::getAttackFactor() const
@@ -2155,7 +2155,7 @@ void Player::death(Creature* _lastHitCreature)
 
 		double deathLossPercent = getLostPercent() * (unfairFightReduction / 100.);
 
-		lostMana = (uint64_t)(sumMana * deathLossPercent);
+		lostMana = static_cast<uint64_t>(sumMana * deathLossPercent);
 
 		while (lostMana > manaSpent && magLevel > 0) {
 			lostMana -= manaSpent;
@@ -2181,7 +2181,7 @@ void Player::death(Creature* _lastHitCreature)
 
 			sumSkillTries += skills[i][SKILLVALUE_TRIES];
 
-			uint32_t lostSkillTries = (uint32_t)(sumSkillTries * deathLossPercent);
+			uint32_t lostSkillTries = static_cast<uint32_t>(sumSkillTries * deathLossPercent);
 			while (lostSkillTries > skills[i][SKILLVALUE_TRIES]) {
 				lostSkillTries -= skills[i][SKILLVALUE_TRIES];
 
@@ -2201,7 +2201,7 @@ void Player::death(Creature* _lastHitCreature)
 		}
 
 		//Level loss
-		uint64_t expLoss = (uint64_t)(experience * deathLossPercent);
+		uint64_t expLoss = static_cast<uint64_t>(experience * deathLossPercent);
 		g_events->eventPlayerOnLoseExperience(this, expLoss);
 
 		if (expLoss != 0) {
@@ -3026,7 +3026,7 @@ void Player::__removeThing(Thing* thing, uint32_t count)
 			item->setParent(nullptr);
 			inventory[index] = nullptr;
 		} else {
-			uint8_t newCount = (uint8_t)std::max<int32_t>(0, item->getItemCount() - count);
+			uint8_t newCount = static_cast<uint8_t>(std::max<int32_t>(0, item->getItemCount() - count));
 			item->setItemCount(newCount);
 
 			//send change to client
@@ -3444,7 +3444,7 @@ uint64_t Player::getGainedExperience(Creature* attacker) const
 {
 	if (g_config.getBoolean(ConfigManager::EXPERIENCE_FROM_PLAYERS)) {
 		Player* attackerPlayer = attacker->getPlayer();
-		if (attackerPlayer && attackerPlayer != this && skillLoss && std::abs((int32_t)(attackerPlayer->getLevel() - level)) <= g_config.getNumber(ConfigManager::EXP_FROM_PLAYERS_LEVEL_RANGE)) {
+		if (attackerPlayer && attackerPlayer != this && skillLoss && std::abs(static_cast<int32_t>(attackerPlayer->getLevel() - level)) <= g_config.getNumber(ConfigManager::EXP_FROM_PLAYERS_LEVEL_RANGE)) {
 			return std::max<uint64_t>(0, std::floor(getLostExperience() * getDamageRatio(attacker) * 0.75));
 		}
 	}
@@ -4090,7 +4090,7 @@ void Player::checkSkullTicks(int32_t ticks)
 
 bool Player::isPromoted() const
 {
-	int32_t promotedVocation = g_vocations.getPromotedVocation(vocation->getId());
+	uint16_t promotedVocation = g_vocations.getPromotedVocation(vocation->getId());
 	return promotedVocation == 0 && vocation->getId() != promotedVocation;
 }
 
@@ -4106,7 +4106,7 @@ double Player::getLostPercent() const
 			lossPercent -= 3;
 		}
 
-		lossPercent -= (int32_t)bitset.count();
+		lossPercent -= static_cast<int32_t>(bitset.count());
 		return std::max<int32_t>(0, lossPercent) / (double)100;
 	} else {
 		double lossPercent;
@@ -4122,7 +4122,7 @@ double Player::getLostPercent() const
 			lossPercent *= 0.7;
 		}
 
-		return lossPercent * pow(0.92, (int32_t)bitset.count()) / 100;
+		return lossPercent * pow(0.92, static_cast<int32_t>(bitset.count())) / 100;
 	}
 }
